Implements the Gaussian blur filter.
This commit implements the `GaussianBlurFilter` class where the shaders used in this class are ported from GPUImage's `GPUImageGaussianBlurFilter` class.

Because this is the first filter implemented in `glfc`. This commit also demonstrates the design pattern of this library. It includes a `Program` class that is responsible for rendering, and the `Filter` class utilizes the `Framebuffer` class to keep intermediate rendering before actually rendering to the binded framebuffer.

The `Filter` is the base class of all filter implementation. It comes with a default one-pass `Render()` method. However, subclasses can override this method for flexibility. For example, the `GaussianBlurFilter` class overrides this method to use the two-pass technique for applying the Gaussian blur effect.