Adds support for device pixel ratio.
This commit updates the `Filter::Render()` method to accept an additional `device_pixel_ratio` argument to make it possible to generate the same filter effect across different screen scale factors.

The `GaussianBlurFilter` takes advantage of this change to calculate the real blur radius and sigma used in the shaders by multiplying the corresponded property values by the `device_pixel_ratio`.