Makes `Filter::Render()` respect the configured blending options.
This commit changes filters to respect the configured blending options before calling `Filter::Render()`. However, the actual fix happens in the `Framebuffer` class.

The `Framebuffer::Bind()` method now enables `GL_BLEND` explicitly so we can use `glBlendFunc(GL_ONE, GL_ZERO)` for the `UpdateTexture()` method. Besides, the original blending options are saved in this method and will be restored in the `Unbind()` method.

As a result, the preconfigured blending options won't affect the process of rendering intermediate filter results in the framebuffer used in `Filter::Render()`.