Fixes memory overflow in Gaussian blur filter's shaders.
This commit fixes the issue that memory overflow may happen to the `standard_gaussian_weights` buffer in `GaussianBlurFilter`'s shader methods.

This issue is caused by the loops that utilize the `kTruekNumberOfOptimizedOffsets` variable in the `GetFragmentShader()` method and the `kNumberOfOptimizedOffsets` variable in the `GetVertexShader()` method.

This bug comes from the `GPUImage` library and is fixed by allocating `kBlurRadius + 2` elements to the `standard_gaussian_weights` buffer instead of just `kBlurRadius + 1`. However, the additional element is set to 0 explicitly to avoid unexpected results.